refactor(jobs): replace container with mediator
Closes DCOS-38202
