+++ src/js/utils/JobUtil.d.ts
@@ -1,3 +1,4 @@
+import { JobData } from "#SRC/js/events/MetronomeClient";
 import Job from "../structs/Job";
 
 export function createJobFromFormModel(
@@ -5,4 +6,4 @@ export function createJobFromFormModel(
   spec?: object
 ): Job;
 export function createFormModelFromSchema(schema: object, job?: Job): object;
-export function createJobSpecFromJob(job: Job): object;
+export function createJobSpecFromJob(job: Job): JobData;
